2019年06月08日 星期六
1. 今天把修炼界面剩下的部分完成了
2. 添加了drop down menu的简单版本
3. 修炼界面中增加了过滤功能
4. 日志系统会计算修炼中的内容
5. 修复了代码生成器里的bug
6. 游戏的时间已经会正常的流逝了

下一步要做的：
1. 商店系统
2. 个人资料

+++ ResourcePackager/ResourcePackager/Data/DataSchema.cpp
@@ -156,7 +156,7 @@ void DataSchema::addValueIntoBuffer(std::unique_ptr<bb::ByteBuffer> &buffer, con
       buffer->putDouble(atof(value.c_str()));
       break;
     case FLOAT:
-      buffer->putLong(atof(value.c_str()));
+      buffer->putFloat(atof(value.c_str()));
       break;
     case FRIEND_ID_SET:
     case FRIEND_ID_VECTOR: {
